Ignore invalid HTTP headers when creating PSR7 objects
